+++ panther/__init__.py
@@ -1,6 +1,6 @@
 from panther.main import Panther  # noqa: F401
 
-__version__ = '3.2.3'
+__version__ = '3.5.0'
 
 
 def version():