+++ panther/_load_configs.py
@@ -11,6 +11,7 @@ from pydantic._internal._model_construction import ModelMetaclass
 from panther._utils import import_class
 from panther.configs import JWTConfig, config
 from panther.exceptions import PantherException
+from panther.middlewares.base import WebsocketMiddleware, HTTPMiddleware
 from panther.routings import finalize_urls, flatten_urls
 from panther.throttling import Throttling
 
@@ -28,6 +29,7 @@ __all__ = (
     'load_jwt_config',
     'load_startup',
     'load_shutdown',
+    'load_auto_reformat',
     'collect_all_models',
     'load_urls',
     'load_panel_urls',
@@ -72,24 +74,43 @@ def load_default_cache_exp(configs: dict, /) -> timedelta | None:
     return configs.get('DEFAULT_CACHE_EXP', config['default_cache_exp'])
 
 
-def load_middlewares(configs: dict, /) -> list:
-    """Collect The Middlewares & Set db_engine If One Of Middlewares Was For DB"""
+def load_middlewares(configs: dict, /) -> dict:
+    """
+    Collect The Middlewares & Set db_engine If One Of Middlewares Was For DB
+    And Return a dict with two list, http and ws middlewares"""
     from panther.middlewares import BaseMiddleware
 
-    middlewares = []
+    middlewares = {'http': [], 'ws': []}
 
-    for path, data in configs.get('MIDDLEWARES', []):
+    for middleware in configs.get('MIDDLEWARES') or []:
+        if not isinstance(middleware, list | tuple):
+            raise _exception_handler(field='MIDDLEWARES', error=f'{middleware} should have 2 part: (path, kwargs)')
+
+        if len(middleware) == 1:
+            path = middleware[0]
+            data = {}
+
+        elif len(middleware) > 2:
+            raise _exception_handler(field='MIDDLEWARES', error=f'{middleware} too many arguments')
+
+        else:
+            path, data = middleware
+
         if path.find('panther.middlewares.db.DatabaseMiddleware') != -1:
             config['db_engine'] = data['url'].split(':')[0]
         try:
             Middleware = import_class(path)  # noqa: N806
-        except AttributeError:
+        except (AttributeError, ModuleNotFoundError):
             raise _exception_handler(field='MIDDLEWARES', error=f'{path} is not a valid middleware path')
 
-        if not issubclass(Middleware, BaseMiddleware):
+        if issubclass(Middleware, BaseMiddleware) is False:
             raise _exception_handler(field='MIDDLEWARES', error='is not a sub class of BaseMiddleware')
 
-        middlewares.append(Middleware(**data))  # noqa: Py Argument List
+        middleware_instance = Middleware(**data)
+        if isinstance(middleware_instance, BaseMiddleware | HTTPMiddleware):
+            middlewares['http'].append(middleware_instance)
+        if isinstance(middleware_instance, BaseMiddleware | WebsocketMiddleware):
+            middlewares['ws'].append(middleware_instance)
     return middlewares
 
 
@@ -118,6 +139,10 @@ def load_startup(configs: dict, /) -> Callable:
 
 def load_shutdown(configs: dict, /) -> Callable:
     return configs.get('SHUTDOWN') and import_class(configs['SHUTDOWN'])
+
+
+def load_auto_reformat(configs: dict, /) -> bool:
+    return configs.get('AUTO_REFORMAT', config['auto_reformat'])
 
 
 def collect_all_models() -> list[dict]: