+++ panther/websocket.py
@@ -31,15 +31,6 @@ class GenericWebsocket(Websocket):
         return await super().send(data=data)
 
 
-def _publish_to_ws_channel(connection_id: str, action: Literal['send', 'close'], data: any):
-    from panther.db.connection import redis
-
-    assert redis.is_connected, 'Redis Is Not Connected.'
-
-    p_data = json.dumps({'connection_id': connection_id, 'action': action, 'data': data})
-    redis.publish('websocket_connections', p_data)
-
-
 async def send_message_to_websocket(connection_id: str, data: any):
     if redis.is_connected:
         _publish_to_ws_channel(connection_id=connection_id, action='send', data=data)
@@ -60,3 +51,12 @@ async def close_websocket_connection(connection_id: str, code: int = status.WS_1
         websocket_connections: WebsocketConnections = config['websocket_connections']
         if connection := websocket_connections.connections.get(connection_id):
             await connection.close(code=code, reason=reason)
+
+
+def _publish_to_ws_channel(connection_id: str, action: Literal['send', 'close'], data: any):
+    from panther.db.connection import redis
+
+    assert redis.is_connected, 'Redis Is Not Connected.'
+
+    p_data = json.dumps({'connection_id': connection_id, 'action': action, 'data': data})
+    redis.publish('websocket_connections', p_data)